+++ ChangeLog.md
@@ -11,7 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- added ...
+- added function 'io/slurp-stream'
+- added try resources block 'try-with'
 
 ### fixed